\name{BePrC}
\alias{BePrC}
\title{Probability matrix for c}
\description{Computes an estimate for the pdf of the discrete variable \code{c_k}, k = 0, 1, ..., K.}
\usage{BePrC(alpha, beta, Pi.r, u.r, epsilon)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }