\name{BeTao}
\alias{BeTao}
\title{tao partition}
\description{Creates a partition for the time axis.}
\usage{BeTao(times)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }