\name{BeUpdU}
\alias{BeUpdU}
\title{u.r update}
\description{Updates discrete variable u at iteration r.}
\usage{BeUpdU(alpha, beta, c.r, Pi.r)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }