% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGPHazard-package.R
\docType{package}
\name{BGPHazard}
\alias{BGPHazard}
\alias{BGPHazard-package}
\title{BGPHazard: A package bayesian nonparametric inference in survival analysis.}
\description{
The BGPHazard package provides three categories of important functions:
simulating, diagnostic and result.
}
\section{Simulating functions}{

The simulating functions are used to make posterior inference for the bayesian survival 
semiparametric models as described by  
Nieto-Barajas and Walker (2002), Nieto-Barajas (2003) and Nieto-Barajas, L. E., & Yin, G. (2008)
}

\section{Diagnostic functions}{

The diagnostic functions are used to make convergence diagnosics plots about the simulations of the parameters/variables.
}

\section{Result functions}{

The result functions are used to produce estimators plots of the hazard function 
along with the survival function defined by the model.
}

