\name{CGaPrU}
\alias{CGaPrU}
\title{Probability matrix for u}
\description{Computes an estimate for the pdf of the discrete variable u_k. k = 0, 1, ..., K.}
\usage{CGaPrU(alpha, beta, c.r, lambda.r)}
\references{
1. Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.

2. Nieto-Barajas, L. E. (2003). Discrete time Markov gamma processes and time dependent covariates in survival analysis. Bulletin of the International Statistical Institute 54th Session. Berlin. (CD-ROM)}
\keyword{ internal }