\name{CGaTao}
\alias{CGaTao}
\title{tao partition}
\description{Creates a partition of the time axis.}
\usage{CGaTao(times, delta, type.t, K)}
\references{
1. Nieto-Barajas, L. E. (2003). Discrete time Markov gamma processes and time dependent covariates in survival analysis. Bulletin of the International Statistical Institute 54th Session. Berlin. (CD-ROM).

2. Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }