\name{CGaUpdC}
\alias{CGaUpdC}
\title{c.r update}
\description{Updates discrete variable c at iteration r.}
\usage{CGaUpdC(alpha, beta, c.r, lambda.r, u.r, epsilon)}
\references{
1. Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.

2. Nieto-Barajas, L. E. (2003). Discrete time Markov gamma processes and time dependent covariates in survival analysis. Bulletin of the International Statistical Institute 54th Session. Berlin. (CD-ROM)}
\keyword{ internal }