\name{CLambdaSumm}
\alias{CLambdaSumm}
\title{Summary for Lambda}
\description{Computes a summary of the hazard and survival function estimated by the chain generated by the gamma process.}
\usage{CLambdaSumm(M, confidence = 0.95)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\value{
\item{SUM.h}{Numeric matrix. Returns an estimate for the hazard rate.}
\item{SUM.S}{Numeric matrix. Returns an estimate for the survival distribution function.}
\item{SUM.H}{Numeric matrix. Returns an estimate fot the cummulative hazard function.}
}
\keyword{ internal }
