\name{GaPrU}
\alias{GaPrU}
\title{Probability matrix for u}
\description{Computes an estimate for the pdf of the discrete variable u_k. k=0,1,...K}
\usage{GaPrU(alpha, beta, c.r, lambda.r)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }