\name{GaUpdC}
\alias{GaUpdC}
\title{c.r update}
\description{Updates discrete variable c at iteration r.}
\usage{GaUpdC(alpha, beta, c.r, lambda.r, u.r, epsilon)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }