\name{GaUpdU}
\alias{GaUpdU}
\title{u.r update}
\description{Updates discrete variable u at iteration r.}
\usage{GaUpdU(alpha, beta, c.r, lambda.r)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }