\name{UpdPi}
\alias{UpdPi}
\title{Pi update}
\description{Updates Pi at iteration r.}
\usage{UpdPi(alpha, beta, c.r, u.r, n, m)}
\references{Nieto-Barajas, L. E. & Walker, S. G. (2002). Markov beta and gamma processes for modelling hazard rates. Scandinavian Journal of Statistics 29, 413-424.}
\keyword{ internal }